src/decoding/batch_decoder.py
import rlp
from typing import Dict, List, Any
import logging
from eth_utils import to_hex, to_bytes
from web3 import Web3

class BatchDecoder:
    def __init__(self):
        self.logger = logging.getLogger(__name__)
        self.logger.setLevel(logging.INFO)
        self.w3 = Web3()

    def decode_rlp(self, data: bytes) -> List[Any]:
        """Decode RLP-encoded data."""
        try:
            return rlp.decode(data)
        except Exception as e:
            self.logger.error(f"RLP decoding error: {str(e)}")
            raise

    def decode_batch(self, batch_data: Dict) -> Dict:
        """Decode a complete batch submission."""
        try:
            # Extract the blob data from the transaction
            blob_data = batch_data['raw_data'].get('blobVersionedHashes', [])
            if not blob_data:
                raise ValueError("No blob data found in transaction")

            decoded_batch = {
                'block_number': batch_data['block_number'],
                'transaction_hash': batch_data['transaction_hash'],
                'blobs': []
            }

            for blob_hash in blob_data:
                # Here we would need to fetch the actual blob data
                # This is a placeholder for the actual blob fetching logic
                blob_content = self._fetch_blob_content(blob_hash)
                if blob_content:
                    decoded_blob = self._decode_blob(blob_content)
                    decoded_batch['blobs'].append(decoded_blob)

            return decoded_batch

        except Exception as e:
            self.logger.error(f"Batch decoding error: {str(e)}")
            raise

    def _fetch_blob_content(self, blob_hash: str) -> bytes:
        """
        Fetch the actual blob content from the network.
        This is a placeholder - actual implementation would depend on the
        specific method used to store and retrieve blob data.
        """
        # TODO: Implement actual blob fetching logic
        return b''

    def _decode_blob(self, blob_content: bytes) -> Dict:
        """
        Decode a single blob into its components.
        This is a placeholder - actual implementation would depend on the
        specific format of the blob data.
        """
        try:
            # Attempt to decode the blob content
            decoded = self.decode_rlp(blob_content)
            
            # Parse the decoded data into a structured format
            # This is a simplified example - actual implementation would be more complex
            return {
                'raw_data': to_hex(blob_content),
                'decoded_data': [to_hex(item) for item in decoded],
                'size': len(blob_content)
            }
        except Exception as e:
            self.logger.error(f"Blob decoding error: {str(e)}")
            raise

    def extract_l2_transactions(self, decoded_batch: Dict) -> List[Dict]:
        """
        Extract individual L2 transactions from a decoded batch.
        This is a placeholder - actual implementation would depend on the
        specific format of the L2 transactions within the batch.
        """
        transactions = []
        for blob in decoded_batch['blobs']:
            # TODO: Implement actual L2 transaction extraction logic
            # This would involve parsing the decoded blob data to find
            # and extract individual L2 transactions
            pass
        return transactions

    def validate_batch(self, decoded_batch: Dict) -> bool:
        """
        Validate the structure and content of a decoded batch.
        This is a placeholder - actual implementation would depend on the
        specific validation requirements.
        """
        try:
            required_fields = ['block_number', 'transaction_hash', 'blobs']
            if not all(field in decoded_batch for field in required_fields):
                return False

            if not isinstance(decoded_batch['blobs'], list):
                return False

            return True
        except Exception as e:
            self.logger.error(f"Batch validation error: {str(e)}")
            return False 