requirements.txt
web3==6.15.1
pandas==2.2.1
numpy==1.26.4
python-dotenv==1.0.1
requests==2.31.0
eth-abi==5.0.1
eth-typing==4.0.0
eth-utils==2.3.1
pytest==8.0.2
black==24.2.0
flake8==7.0.0
plotly==5.19.0
matplotlib==3.8.3
sqlalchemy==2.0.27
psycopg2-binary==2.9.9 